# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API Documentation  for More-Recipes. More-Recipes is a recipe sharing  and social connecting platform where users can show  their cuisine expertise by posting recipes for other  users to view and learn from",
    "version": "v1",
    "title": "More-Recipes API",
    "contact": {
      "email": "larrystonegroup@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/swagger-api/swagger-ui/blob/master/LICENSE"
    }
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users/changepassword": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Allows an authenticated user change password",
        "description": "By providing valid login credentials a user is able to change his/her password",
        "operationId": "changePassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "The oldPassword and newPassword details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password Changed Successfully",
            "schema": {
              "$ref": "#/definitions/ChangePasswordResponse"
            }
          },
          "400": {
            "description": "Validation exception",
            "schema": {
              "$ref": "#/definitions/ChangePasswordError"
            }
          },
          "401": {
            "description": "Incorrect Password",
            "schema": {
              "$ref": "#/definitions/SigninError"
            }
          },
          "default": {
            "description": "Error Changing password"
          }
        }
      }
    },
    "/users/{userId}/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get an authenticated user profile details",
        "description": "This method is called to fetch a user in the system",
        "operationId": "getUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/UserNotFound"
            }
          },
          "default": {
            "description": "Unable to update user details"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Allows an authenticated user update his/her profile details",
        "description": "By providing valid login credentials a user is able to modify his/her profile details",
        "operationId": "modifyUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Form input model",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }],
        "responses": {
          "200": {
            "description": "User record updated",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "Validation exception",
            "schema": {
              "$ref": "#/definitions/SignupValidation"
            }
          },
          "503": {
            "description": "Image upload error"
          },
          "409": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/SignupValidation"
            }
          },
          "default": {
            "description": "Error updating profile"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserResponse": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "john doe"
        },
        "username": {
          "type": "string",
          "example": "johndoe"
        },
        "email": {
          "type": "string",
          "example": "johndoe@company.com"
        },
        "imageUrl": {
          "type": "string",
          "example": "http://ssfsfnsfnsnsfsfs.png"
        }
      }
    },
...
    "SignupValidation": {
      "type": "object",
      "required": [
        "success",
        "message"
      ],
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "Username already taken!"
        }
      }
    },
...
    "SigninError": {
      "type": "object",
      "required": [
        "success",
        "message"
      ],
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "Invalid Login Credentials!"
        }
      }
    },
    "UserNotFound": {
      "type": "object",
      "required": [
        "success",
        "message"
      ],
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "Username not found!"
        }
      }
    },
    "ChangePassword": {
      "type": "object",
      "required": [
        "oldPassword",
        "newPassword"
      ],
      "properties": {
        "oldPassword": {
          "type": "string",
          "example": "oldpassword"
        },
        "newPassword": {
          "type": "string",
          "example": "newpassword"
        }
      }
    },
    "ChangePasswordResponse": {
      "type": "object",
      "required": [
        "message",
        "status"
      ],
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "message": {
          "type": "string",
          "example": "Password changed successfully"
        }
      }
    },
    "ChangePasswordError": {
      "type": "object",
      "required": [
        "message",
        "status"
      ],
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "message": {
          "type": "string",
          "example": "Password must be at least 6 characters!"
        }
      }
    },
...
  }
}